msc: View Log flyout with pop-out window; clamp windows to work area
Log viewer extracted to a shared LogViewerControl (tail, filter,
autoscroll). View Log now opens a flyout anchored to the button by
default; its pop-out button (and Ctrl+L) opens the standalone LogWindow.

LogWindow centered itself with hardcoded 2250x1500 math against a 900x600
window, so on shorter displays Y went negative and it opened above the
visible area. Both windows now center using actual window size offset by
the work area origin and clamped to non-negative coordinates; MainWindow
also caps its initial size to the work area.
