+++ gui/ManagedSoftwareCenter/Controls/LogViewerControl.xaml
@@ -1,0 +1,90 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- LogViewerControl.xaml - Reusable session log viewer (tail + filter).
+     Hosted by the View Log flyout on the Updates page and by the standalone
+     LogWindow. The pop-out button is only shown in flyout mode. -->
+<UserControl
+    x:Class="Cimian.GUI.ManagedSoftwareCenter.Controls.LogViewerControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <!-- Toolbar -->
+        <Grid Grid.Row="0" Padding="12,8">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+
+            <TextBox x:Name="SearchBox"
+                     PlaceholderText="Filter log..."
+                     TextChanged="SearchBox_TextChanged"
+                     MaxWidth="300"
+                     HorizontalAlignment="Left"/>
+
+            <StackPanel Grid.Column="1" Orientation="Horizontal" Spacing="4">
+                <ToggleButton x:Name="AutoScrollToggle"
+                              IsChecked="True"
+                              ToolTipService.ToolTip="Auto-scroll to bottom"
+                              Click="AutoScrollToggle_Click">
+                    <FontIcon Glyph="&#xE74B;" FontSize="14"/>
+                </ToggleButton>
+                <Button Click="Refresh_Click"
+                        ToolTipService.ToolTip="Reload log">
+                    <FontIcon Glyph="&#xE72C;" FontSize="14"/>
+                </Button>
+                <Button Click="Clear_Click"
+                        ToolTipService.ToolTip="Clear display">
+                    <FontIcon Glyph="&#xE894;" FontSize="14"/>
+                </Button>
+                <Button x:Name="PopOutButton"
+                        Visibility="Collapsed"
+                        Click="PopOut_Click"
+                        ToolTipService.ToolTip="Open in separate window">
+                    <FontIcon Glyph="&#xE8A7;" FontSize="14"/>
+                </Button>
+            </StackPanel>
+        </Grid>
+
+        <!-- Log Content -->
+        <ScrollViewer x:Name="LogScrollViewer" Grid.Row="1"
+                      VerticalScrollBarVisibility="Auto"
+                      HorizontalScrollBarVisibility="Auto">
+            <TextBlock x:Name="LogTextBlock"
+                       FontFamily="Cascadia Mono, Consolas, Courier New"
+                       FontSize="12"
+                       IsTextSelectionEnabled="True"
+                       TextWrapping="Wrap"
+                       Padding="12"
+                       Text="Loading log..."/>
+        </ScrollViewer>
+
+        <!-- Status Bar -->
+        <Grid Grid.Row="2" Padding="12,6"
+              Background="{ThemeResource LayerFillColorDefaultBrush}"
+              BorderBrush="{ThemeResource CardStrokeColorDefaultBrush}"
+              BorderThickness="0,1,0,0">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+
+            <TextBlock x:Name="StatusText"
+                       FontSize="11"
+                       Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                       VerticalAlignment="Center"
+                       Text="Watching log file..."/>
+
+            <TextBlock x:Name="LineCountText"
+                       Grid.Column="1"
+                       FontSize="11"
+                       Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                       VerticalAlignment="Center"/>
+        </Grid>
+    </Grid>
+</UserControl>
