+++ gui/ManagedSoftwareCenter/Views/LogWindow.xaml
@@ -3,83 +3,24 @@
     x:Class="Cimian.GUI.ManagedSoftwareCenter.Views.LogWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:controls="using:Cimian.GUI.ManagedSoftwareCenter.Controls"
     Title="Managed Software Update Log">
 
     <Grid Padding="0">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
             <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
 
-        <!-- Integrated Title Bar + Toolbar -->
-        <Grid x:Name="AppTitleBar" Grid.Row="0" Padding="12,8,180,8" Height="48"
+        <!-- Integrated Title Bar (drag region) -->
+        <Grid x:Name="AppTitleBar" Grid.Row="0" Padding="16,0,180,0" Height="40"
               Background="Transparent">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*"/>
-                <ColumnDefinition Width="Auto"/>
-            </Grid.ColumnDefinitions>
-
-            <!-- Search Box -->
-            <TextBox x:Name="SearchBox"
-                     PlaceholderText="Filter log..."
-                     TextChanged="SearchBox_TextChanged"
-                     MaxWidth="300"
-                     HorizontalAlignment="Left"/>
-
-            <!-- Action Buttons -->
-            <StackPanel Grid.Column="1" Orientation="Horizontal" Spacing="4">
-                <ToggleButton x:Name="AutoScrollToggle"
-                              IsChecked="True"
-                              ToolTipService.ToolTip="Auto-scroll to bottom"
-                              Click="AutoScrollToggle_Click">
-                    <FontIcon Glyph="&#xE74B;" FontSize="14"/>
-                </ToggleButton>
-                <Button Click="Refresh_Click"
-                        ToolTipService.ToolTip="Reload log">
-                    <FontIcon Glyph="&#xE72C;" FontSize="14"/>
-                </Button>
-                <Button Click="Clear_Click"
-                        ToolTipService.ToolTip="Clear display">
-                    <FontIcon Glyph="&#xE894;" FontSize="14"/>
-                </Button>
-            </StackPanel>
-        </Grid>
-
-        <!-- Log Content -->
-        <ScrollViewer x:Name="LogScrollViewer" Grid.Row="1"
-                      VerticalScrollBarVisibility="Auto"
-                      HorizontalScrollBarVisibility="Auto">
-            <TextBlock x:Name="LogTextBlock"
-                       FontFamily="Cascadia Mono, Consolas, Courier New"
+            <TextBlock Text="Managed Software Update Log"
                        FontSize="12"
-                       IsTextSelectionEnabled="True"
-                       TextWrapping="Wrap"
-                       Padding="12"
-                       Text="Loading log..."/>
-        </ScrollViewer>
-
-        <!-- Status Bar -->
-        <Grid Grid.Row="2" Padding="12,6"
-              Background="{ThemeResource LayerFillColorDefaultBrush}"
-              BorderBrush="{ThemeResource CardStrokeColorDefaultBrush}"
-              BorderThickness="0,1,0,0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*"/>
-                <ColumnDefinition Width="Auto"/>
-            </Grid.ColumnDefinitions>
-
-            <TextBlock x:Name="StatusText"
-                       FontSize="11"
-                       Foreground="{ThemeResource TextFillColorSecondaryBrush}"
-                       VerticalAlignment="Center"
-                       Text="Watching log file..."/>
-
-            <TextBlock x:Name="LineCountText"
-                       Grid.Column="1"
-                       FontSize="11"
-                       Foreground="{ThemeResource TextFillColorSecondaryBrush}"
                        VerticalAlignment="Center"/>
         </Grid>
+
+        <!-- Log Viewer -->
+        <controls:LogViewerControl x:Name="LogViewer" Grid.Row="1"/>
     </Grid>
 </Window>
