+++ gui/ManagedSoftwareCenter/Views/UpdatesPage.xaml
@@ -4,7 +4,8 @@
     x:Class="Cimian.GUI.ManagedSoftwareCenter.Views.UpdatesPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:converters="using:Cimian.GUI.ManagedSoftwareCenter.Converters">
+    xmlns:converters="using:Cimian.GUI.ManagedSoftwareCenter.Converters"
+    xmlns:controls="using:Cimian.GUI.ManagedSoftwareCenter.Controls">
 
     <Page.Resources>
         <converters:StringFormatConverter x:Key="StringFormatConverter" />
@@ -44,6 +45,24 @@
                                 <FontIcon Glyph="&#xE7C3;" FontSize="16"/>
                                 <TextBlock Text="View Log" Margin="8,0,0,0"/>
                             </StackPanel>
+                            <Button.Flyout>
+                                <Flyout x:Name="LogFlyout"
+                                        Placement="BottomEdgeAlignedRight"
+                                        Opening="LogFlyout_Opening"
+                                        Closed="LogFlyout_Closed">
+                                    <Flyout.FlyoutPresenterStyle>
+                                        <Style TargetType="FlyoutPresenter">
+                                            <Setter Property="MaxWidth" Value="820"/>
+                                            <Setter Property="MaxHeight" Value="560"/>
+                                            <Setter Property="Padding" Value="0"/>
+                                            <Setter Property="CornerRadius" Value="8"/>
+                                        </Style>
+                                    </Flyout.FlyoutPresenterStyle>
+                                    <controls:LogViewerControl x:Name="FlyoutLogViewer"
+                                                               Width="780"
+                                                               Height="520"/>
+                                </Flyout>
+                            </Button.Flyout>
                         </Button>
                         <Button x:Name="InstallNowButton"
                                 Style="{StaticResource AccentButtonStyle}"
